refactor(blackjack) : move enums to domain package
