+++ src/main/kotlin/blackjack/domain/Card.kt
@@ -1,8 +1,5 @@
 package blackjack.domain
 
-import blackjack.enums.CardShape
-import blackjack.enums.CardType
-
 class Card(
     private val shape: CardShape,
     val type: CardType
